pyproject.toml
[build-system]
requires = ["setuptools>=61", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "deployment-agent"
version = "0.1.0"
description = "Autonomous ML model deployment agent (Phase 1: classical ML)"
requires-python = ">=3.11"
dependencies = [
  "langgraph>=0.2.0",
  "langchain-core>=0.3.0",
  "google-genai>=1.0.0",
  "groq>=0.11.0",
  "python-dotenv>=1.0.0",
  "joblib>=1.3.0",
  "scikit-learn>=1.3.0",
  "requests>=2.31.0",
  "fastapi>=0.110.0",
  "uvicorn[standard]>=0.27.0",
  "python-multipart>=0.0.9",
  "onnxruntime>=1.18.0",
  "onnx>=1.16.0",
  "torch>=2.1.0",
  "tensorflow>=2.15.0",
]

[project.scripts]
deploy-agent = "deployment_agent.cli:main"

[tool.setuptools.packages.find]
where = ["src"]
